requirements.txt
matplotlib~=3.9.2
numpy~=1.26.4
torch~=2.4.0
pillow~=10.4.0
torchvision~=0.19.0